removed connection with autoexec.py
since the cleaner now is a service we can omit the autoexec.py altogether. There is now a function that removes any existing call to this addon from autoexec.py. This prevents the addon starting twice after the update to version 2.

also, modified several settings to use sliders, and added a scanning interval option